Caldroid Sample
Caldroid is a fragment that display calendar with dates in a month.
Caldroid can be used as embedded fragment, or as dialog fragment. User
can also swipe left/right to navigate to different months.

Caldroid is fully localized. You can customize start day of the week for
different countries. By default calendar start on Sunday.

https://github.com/roomorama/Caldroid